+++ app/interactors/users/save.rb
@@ -1,19 +1,8 @@
 module Users
   class Save
-    include Interactor
+    include Interactor::Organizer
 
-    delegate :user_params, to: :context
-
-    def call
-      context.user = user
-
-      context.fail! unless user.update(user_params)
-    end
-
-    private
-
-    def user
-      @user ||= context.user || User.new
-    end
+    organize Users::Save::Execute,
+             Users::AccessToken
   end
 end
